requirements.txt
langchain-groq
langchain-ollama
pandas
numpy
python-dotenv
langchain_community
langchain_core
streamlit